datapath/dpdk: reduce the number of TX queues of phy port
Reduce the number of TX queues to the number of enabled lcores instead
of RTE_MAX_LCORE. Assign the TX queue to enabled lcore only, because
the TX queues for disabled lcores will never be used.
This makes the device which has small number of queues usable.

This will enable the use of NICs with a low number of queues.

Change-Id: If21642fa59cfee78faae12ef7be86ece8e6f362a
Signed-off-by: Hiroshi Shimamoto <h-shimamoto@ct.jp.nec.com>
Reviewed-by: Hayato Momma <h-momma@ce.jp.nec.com>
Tested-by: Mark D. Gray <mark.d.gray@intel.com>