lib/dpif-dpdk: fix memory leak in upcall
valgrind detected memory leak in upcall.

==6318== 4,784 bytes in 46 blocks are definitely lost in loss record 403 of 415
==6318==    at 0x4C2845D: malloc (in /usr/lib64/valgrind/vgpreload_memcheck-amd64-linux.so)
==6318==    by 0x4C7037: xmalloc (util.c:108)
==6318==    by 0x4AD4F4: ofpbuf_new_with_headroom (ofpbuf.c:140)
==6318==    by 0x4AD613: ofpbuf_clone_data (ofpbuf.c:209)
==6318==    by 0x4F167F: dpdk_link_recv_packet (dpdk-link.c:298)
==6318==    by 0x4E3FB2: dpif_dpdk_recv (dpif-dpdk.c:1676)
==6318==    by 0x471218: dpif_recv (dpif.c:1333)
==6318==    by 0x452C9C: udpif_dispatcher (ofproto-dpif-upcall.c:838)
==6318==    by 0x2539FDF2: start_thread (in /usr/lib64/libpthread-2.17.so)
==6318==    by 0x25DB801C: clone (in /usr/lib64/libc-2.17.so)

The 'buf' is declared in dpif_dpdk_recv() and newly allocated by
ofpbuf_clone_data(). Only the caller-provided 'buf' is freed,
while the declared 'buf' is not. This leads to memory leak.
The 'upcall->key' and 'upcall->userdata' should be the pointers to the
caller-provided 'buf' as specified for recv() in dpif-provider.h .

The following is the snippet of the specification of recv()
in dpif-provider.h,which was updated in commit 693eb1ff
(Squashed 'openvswitch/' changes from b254093..e8b77d8).
     * The implementation should point 'upcall->key' and 'upcall->userdata'
     * (if any) into data in the caller-provided 'buf'.  The implementation may
     * also use 'buf' for storing the data of 'upcall->packet'.  If necessary
     * to make room, the implementation may reallocate the data in 'buf'.

Also, upcall->packet and caller-provided 'buf' should not refer
to the same ofpbuf memory area, because both are freed in upcall_destroy()
and this leads to double free.

Change-Id: I50402455c296a68386fd711399fb957c5c8e50fc
Signed-off-by: Hiroshi Shimamoto <h-shimamoto@ct.jp.nec.com>
Signed-off-by: Hiroaki Yamamoto <yamamoto.hrk@ncos.nec.co.jp>
Reviewed-by: Hayato Momma <h-momma@ce.jp.nec.com>
Tested-by: Mark D. Gray <mark.d.gray@intel.com>